Add enemy system.

+++ sprite.h
@@ -12,7 +12,7 @@ typedef struct spritehit_t
 } spritehit_t;
 
 HANDLE loadSprite(TEXT* baseName, u8 id);
-u16 projectSprite(const u16 x, const u16 y, spritehit_t* hit, const f16 f_viewCos, const f16 f_viewSin);
+u16 projectSprite(const f16 x, const f16 y, spritehit_t* hit, const f16 f_viewCos, const f16 f_viewSin);
 void drawSprite(const spritehit_t* spriteHit);
 
 #endif
